Add files via upload
style panels to have a background image and a heading. use Flexbox to align everything and then add a little bit of JavaScript so that when we click one of them, it changes the class to active and it then expands with a CSS transition. so will be using JavaScript for the event. Listeners will be using CSS transitions to give it the smooth through the effect. the heading doesnt come in until the card is fully expanded as well and were using just background images and headings, but you could make these into testimonials or FAQs or whatever.