Add files via upload
progress steps you might see these in different kinds of websites with forms multi basically like multi level forms, shopping carts, things like that. no form or anything, but will have these buttons to control the steps. it got blue outline, if click next, then it goes to two. so the line gets transition goes over to two to border lights up, click next again and notice at the beginning. Previous is disabled so there's no reason to click Previous and then we can keep going, and when we click, when we get to the last one, the Next is disabled.