Add files via upload
basically we just have a simple article. this could be anything at all, any website, any blog, whatever. at top we have the little navigation icon. if click this, what's going to happen is it actually rotates the entire website, the entire container, and then it brings in the NAV, the navigation items from the side and the icon also changes. so basically we have two icons open and close. were going to position them so that when we click it, this circle rotates and shows the close button. obviously we need a little bit of JavaScript to be able to add and remove the specific class, which were going to call show NAV and that will rotate and bring the menu in. so its going to be mostly CSS that were dealing with. we need to add specific styles for specific elements. when the show NAV class is present, the only JavaScript we really need is to add or move the class when we click this button.