Add files via upload
a cool little widget that you can add to your website or to your application. its a hidden search, and we just have this button with a little search icon from FontAwesome. as soon as click it, it fires off a JavaScript event that activates a specific class where we add a transition for the width of this input to get wider and the button to move over using a transform. TranslateX() and if click it again its going to shrink back up. so will be using CSS transitions along with just a little bit of JavaScript to add the event and toggle the class.