Add files via upload
a blurry loading page or loading image. the image starts blurred and start with 0%, and then that goes up to 100% and the image comes into focus. so could use this to load just about anything. were just dong a blur effect on the image but could load in your site, you could have any other elements in the DOM load in that specific amount of time from when we go to 0 to 100 and will have to map different number ranges to the 0 to 100 load, for instance, the image blur is going to be 30 pixels when it starts and its going to go down to zero as we go from 0 to 100. there's a StackOverflow function (https://stackoverflow.com/questions/10756313/javascript-jquery-map-a-range-of-numbers-to-another-range-of-numbers) that were going to use to help us do that, to help us map one number range to another.