Add files via upload
a scroll animation, which is is pretty simple. websites where you start scrolling and parts of the site, you know, images or sections of the website start to come in. when start to scroll down, the next one comes in from the left, go down some more, next one comes in from the right and it's going to alternate where it comes in from. using CSS transitions to have it come in smoothly. going to use JavaScript to basically tell where the trigger point is for the next box to come in. then were going to add a specific class to show it. then if go back up, itll remove the class when it hits the trigger point and it'll all slide back out. useful in a lot of different UIs and different websites.

a lot of ppl use jQuery for something like this or a library like animate on scroll, this is just something simple, this is very lightweight, you know, no extra libraries or anything and its really easy to do