Add files via upload
split screen with a background image, with an overlay, with some content inside. if we click on one side, it's going to stretch out to about 75%, the other side, 25%. if hover on the right side, it's going to do the same thing. add a little hover effect onto these. just a nice looking little landing page. very CSS heavy. a little bit of JavaScript at the end for the hover effects, but lots of CSS in this project.